Create Channels folder and move UserChannel inside
