+++ src/fdc3/dotnet/DesktopAgent/test/MorganStanley.ComposeUI.DesktopAgent.Tests/MorganStanley.ComposeUI.DesktopAgent.Tests.csproj
@@ -5,8 +5,8 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <IsPackable>false</IsPackable>
-    <AssemblyName>MorganStanley.ComposeUI.Fdc3.$(MSBuildProjectName)</AssemblyName>
-    <RootNamespace>MorganStanley.ComposeUI.Fdc3.$(MSBuildProjectName.Replace(" ", "_"))</RootNamespace>
+    <AssemblyName>MorganStanley.ComposeUI.Fdc3.DesktopAgent.Tests</AssemblyName>
+    <RootNamespace>MorganStanley.ComposeUI.Fdc3.DesktopAgent.Tests</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
